+++ .env
@@ -1,0 +1,3 @@
+CLIENT_ID=1fe1ac6261a105f3ee49
+CLIENT_SECRET=1bb7b9145181033a592d38630ef3591880ae096a
+APP_SECRET=DEFAULT
