+++ app/index.html
@@ -6,11 +6,7 @@
   <h1>SUCH WOW WEBPAGE</h1>
 </head>
 <body>
-  <div id ="root">
-    <!-- <section id="currentAssignments"></section>
-    <section id="futureAssignments"></section>
-    <section id="previousAssignments"></section> -->
-  </div>
+  <div id ="root"></div>
   <script type="text/javascript" src="bundle.js"></script>
 </body>
 </html>
