+++ app/js/components/assignment_box/assignment_box.jsx
