+++ app/js/components/assignment_box/assignment_header/assignment_header.jsx
