+++ app/js/components/assignment_box/assignment_list/assignment/assignment.jsx
