+++ app/js/components/assignment_box/assignment_list/assignment_list.jsx
