+++ app/js/constants/action_types.js
@@ -1,3 +1,7 @@
+//login actions
+module.exports.USER_SIGNUP = 'USER_SIGNUP';
+module.exports.USER_LOGIN = 'USER_LOGIN';
+
 // network actions
 module.exports.REQUEST_ASSIGNMENTS = 'REQUEST_ASSIGNMENTS';
 module.exports.RECEIVE_ASSIGNMENTS = 'RECEIVE_ASSIGNMENTS';
