+++ app/js/containers/dashboard.jsx
@@ -1,11 +1,8 @@
 var React = require('react');
 var ReactDOM = require('react-dom');
-var connect = require('react-redux').connect;
 var AssignmentBox = require(__dirname + '/../components/assignment_box/assignment_box.jsx');
-var bindActionCreators = require('redux').bindActionCreators;
-var assignmentsActions = require(__dirname + '/../actions/assignments_actions');
 
-var Dashboard = React.createClass({
+var Dashboard = module.exports = React.createClass({
   render: function() {
     return (
       <div>
@@ -17,17 +14,3 @@ var Dashboard = React.createClass({
   }
 });
 
-function mapStateToProps(state) {
-  return { assignments: state.assignments };
-};
-
-function mapDispatchToProps(dispatch) {
-  return bindActionCreators(assignmentsActions, dispatch);
-};
-
-var App = connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(Dashboard);
-
-module.exports = App;
